+++ src/attachment.rs
@@ -1,11 +1,11 @@
 use slack::SlackText;
-use error::Result;
-use hex::{HexColor, HexColorT};
-use helper::opt_str_to_slacktext;
+use error::{Error, Result};
+use hex::HexColor;
+use TryInto;
 
 /// Slack allows for attachments to be added to messages. See
 /// https://api.slack.com/docs/attachments for more information.
-#[derive(RustcEncodable, Debug)]
+#[derive(RustcEncodable, Debug, Default)]
 pub struct Attachment {
     /// Required text for attachment.
     /// Slack will use this text to display on devices that don't support markup.
@@ -14,68 +14,11 @@ pub struct Attachment {
     pub text: Option<SlackText>,
     /// Optional text that appears above attachment
     pub pretext: Option<SlackText>,
-    /// Color of attachment
-    pub color: HexColor,
+    /// Optional color of attachment
+    pub color: Option<HexColor>,
     /// Fields are defined as an array, and hashes contained within it will be
     /// displayed in a table inside the message attachment.
     pub fields: Option<Vec<Field>>,
-}
-
-/// Attachment template to simplify constructing attachments
-/// for common use cases.
-#[derive(Debug)]
-pub enum AttachmentTemplate<'a> {
-    /// Specify all attributes of attachment
-    Complete {
-        /// Required text for attachment.
-        /// Slack will use this text to display on devices that don't support markup.
-        fallback: &'a str,
-        /// Optional primary text of attachment
-        text: Option<&'a str>,
-        /// Optional text that appears above attachment
-        pretext: Option<&'a str>,
-        /// Color string can be any hex code starting with #
-        color: &'a str,
-        /// Fields are defined as an array, and hashes contained within it will
-        /// be displayed in a table inside the message attachment.
-        fields: Option<Vec<Field>>,
-    },
-    /// Provide only text and color for attachment
-    /// other values will be defaulted
-    Text {
-        /// Text to send
-        text: &'a str,
-        /// Color string can be any hex code starting with #
-        color: &'a str,
-    },
-}
-
-impl Attachment {
-    /// Construct new attachment based on template provided
-    pub fn new(t: AttachmentTemplate) -> Result<Attachment> {
-        match t {
-            AttachmentTemplate::Complete { fallback, text, pretext, color, fields } => {
-                let c = try!(HexColorT::new(color));
-                Ok(Attachment {
-                    fallback: SlackText::new(fallback),
-                    text: opt_str_to_slacktext(&text),
-                    pretext: opt_str_to_slacktext(&pretext),
-                    color: c,
-                    fields: fields,
-                })
-            }
-            AttachmentTemplate::Text { text, color } => {
-                let c = try!(HexColorT::new(color));
-                Ok(Attachment {
-                    fallback: SlackText::new(text),
-                    text: Some(SlackText::new(text)),
-                    pretext: None,
-                    color: c,
-                    fields: None,
-                })
-            }
-        }
-    }
 }
 
 /// Fields are defined as an array, and hashes contained within it will
@@ -95,11 +38,104 @@ pub struct Field {
 
 impl Field {
     /// Construct a new field
-    pub fn new(title: &str, value: &str, short: Option<bool>) -> Field {
+    pub fn new<S: Into<String>, ST: Into<SlackText>>(title: S,
+                                                     value: ST,
+                                                     short: Option<bool>)
+                                                     -> Field {
         Field {
-            title: title.to_owned(),
-            value: SlackText::new(value),
+            title: title.into(),
+            value: value.into(),
             short: short,
+        }
+    }
+}
+
+/// `AttachmentBuilder` is used to build a `Attachment`
+#[derive(Debug)]
+pub struct AttachmentBuilder {
+    inner: Result<Attachment>,
+}
+
+impl AttachmentBuilder {
+    /// Make a new `AttachmentBuilder`
+    ///
+    /// Fallback is the only required field which is a plain-text summary of the attachment.
+    pub fn new<S: Into<SlackText>>(fallback: S) -> AttachmentBuilder {
+        AttachmentBuilder {
+            inner: Ok(Attachment { fallback: fallback.into(), ..Default::default() }),
+        }
+    }
+
+    /// Optional text that appears within the attachment
+    pub fn text<S: Into<SlackText>>(self, text: S) -> AttachmentBuilder {
+        match self.inner {
+            Ok(mut inner) => {
+                inner.text = Some(text.into());
+                AttachmentBuilder { inner: Ok(inner) }
+            }
+            _ => self,
+        }
+    }
+
+    /// Set the color of the attachment
+    ///
+    /// The color can be one of:
+    ///
+    /// 1. The built-in `SlackColor` variants: `SlackColor::Good`, etc.
+    /// 1. Any valid color `String`: good`, `warning`, `danger`
+    /// 3. Any valid hex color code: `#b13d41`
+    ///
+    /// hex color codes will be checked to ensure a valid hex number is provided
+    pub fn color<C: TryInto<HexColor, Err = Error>>(self, color: C) -> AttachmentBuilder {
+        match self.inner {
+            Ok(mut inner) => {
+                match color.try_into() {
+                    Ok(c) => {
+                        inner.color = Some(c);
+                        AttachmentBuilder { inner: Ok(inner) }
+                    }
+                    Err(e) => AttachmentBuilder { inner: Err(e) },
+                }
+            }
+            _ => self,
+        }
+    }
+
+    /// Optional text that appears above the attachment block
+    pub fn pretext<S: Into<SlackText>>(self, pretext: S) -> AttachmentBuilder {
+        match self.inner {
+            Ok(mut inner) => {
+                inner.pretext = Some(pretext.into());
+                AttachmentBuilder { inner: Ok(inner) }
+            }
+            _ => self,
+        }
+    }
+
+    /// Fields are defined as an array, and hashes contained within it will be
+    /// displayed in a table inside the message attachment.
+    pub fn fields(self, fields: Vec<Field>) -> AttachmentBuilder {
+        match self.inner {
+            Ok(mut inner) => {
+                inner.fields = Some(fields);
+                AttachmentBuilder { inner: Ok(inner) }
+            }
+            _ => self,
+        }
+    }
+
+    /// Attempt to build the `Attachment`
+    pub fn build(self) -> Result<Attachment> {
+        // set text to equal fallback if text wasn't specified
+        match self.inner {
+            Ok(mut inner) => {
+                if inner.text.is_none() {
+                    inner.text = Some(inner.fallback.clone())
+                }
+                Ok(inner)
+            }
+            _ => self.inner,
+
         }
     }
 }