+++ src/lib.rs
@@ -30,14 +30,10 @@ extern crate rustc_serialize;
 extern crate quick_error;
 
 pub use slack::{Slack, SlackText, SlackLink};
-
-pub use payload::{Payload, PayloadTemplate};
-
-pub use attachment::{Attachment, AttachmentTemplate, Field};
-
-pub use error::{Error, Result};
-
+pub use payload::{Payload, PayloadBuilder};
+pub use attachment::{Attachment, AttachmentBuilder, Field};
 pub use hex::{SlackColor, HexColor};
+pub use error::{Error, Result};
 
 mod helper;
 mod error;
@@ -45,3 +41,41 @@ mod hex;
 mod payload;
 mod attachment;
 mod slack;
+
+/// Waiting to stabilize: https://github.com/rust-lang/rust/issues/33417
+///
+/// An attempted conversion that consumes `self`, which may or may not be expensive.
+///
+/// Library authors should not directly implement this trait, but should prefer implementing
+/// the [`TryFrom`] trait, which offers greater flexibility and provides an equivalent `TryInto`
+/// implementation for free, thanks to a blanket implementation in the standard library.
+///
+/// [`TryFrom`]: trait.TryFrom.html
+pub trait TryInto<T>: Sized {
+    /// The type returned in the event of a conversion error.
+    type Err;
+
+    /// Performs the conversion.
+    fn try_into(self) -> ::std::result::Result<T, Self::Err>;
+}
+
+/// Waiting to stabilize: https://github.com/rust-lang/rust/issues/33417
+///
+/// Attempt to construct `Self` via a conversion.
+pub trait TryFrom<T>: Sized {
+    /// The type returned in the event of a conversion error.
+    type Err;
+
+    /// Performs the conversion.
+    fn try_from(T) -> ::std::result::Result<Self, Self::Err>;
+}
+
+impl<T, U> TryInto<U> for T
+    where U: TryFrom<T>
+{
+    type Err = U::Err;
+
+    fn try_into(self) -> ::std::result::Result<U, U::Err> {
+        U::try_from(self)
+    }
+}