+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 DEPRECATIONS/BREAKING CHANGES:
 
+ - `Payload`
+  - `PayloadTemplate` is removed.
+  - `unfurl_links` is now bool. No integer transformation is required by the slack API anymore.
+ - `Attachment`
+  - `AttachmentTemplate` is removed.
+  - `color` is now optional.
  - `SlackError` is now `Error` and `SlackResult` is now `Result`. In addition, `ErrorKind` has been
  removed. All `Error` variants have also been renamed, removing the superfluous `Err` prefix:
   - `ErrSlackResp` -> `Slack`
@@ -10,3 +16,20 @@ DEPRECATIONS/BREAKING CHANGES:
   - `ErrHexColor` -> `HexColor`
   - `ErrEncoder` -> `Encoder`
   - `ErrCurl` -> `Curl`
+ - `SlackLink`, `SlackText`
+  - The `Display` trait is now used to format strings for sending to slack rather than `Debug`.
+  The `Debug` impl is derived now.
+
+FEATURES
+
+ - `PayloadBuilder` and `AttachmentBuilder` should be used for building a `Payload` or `Attachment`
+ respectively. Errors won't be returned until the final `build` function is called. At this point,
+ only the first error is displayed. Subsequent errors will only appear once the first error is
+ fixed.
+ - `Payload`
+  - Added `unfurl_media` option
+
+OTHER
+ - `TryFrom` and `TryInto` traits have been added temporarily to this crate until they are
+ formalized in rust proper. See https://github.com/rust-lang/rust/issues/33417 for details.
+ - `HexColorT` trait removed. Conversions are used instead.